refactor: Add createUser method to userAPI module
This commit adds a new method, createUser, to the userAPI module. The createUser method is responsible for making a POST request to create a new user. It takes in user data as a parameter and sends it to the server. This method is used in the SignUp component to handle form submission and create a new user.

The createUser method is implemented using the axiosClient.post function to send the request to the "/users" endpoint. The response from the server is logged to the console. If the response status is "success", a success notification is displayed and the user is redirected to the sign-in page. If there is an error, an error notification is displayed.

This commit also includes the addition of the NotiAlert component to display notifications. The NotiAlert component is used to show success or error messages when creating a user.